added data access interface package and renamed data access package to data access object
