+++ src/main/java/use_case/login/LoginInteractor.java
@@ -1,6 +1,5 @@
 package use_case.login;
 
-import data_access.FileAccessDataObject;
 import entity.Player;
 
 import java.io.IOException;
